(feature) Plugin dependency check
Some of our themes have a dependency on one or more plugins. Most of the
time it isn't documented anywhere and can cause confusion when new
developers pick up the project.

This commit adds a plugin dependency check where developers can define
the plugins required to run the theme in `di.php`

The checks will be ran when the theme is activated and any plugins that
are required and not active will have a 'warning' message displayed.