Upgrade to WP_Mock v0.3.0
The latest release is 0.4.2, but 0.4 and up requires v7.0 of PHPUnit, which required PHP v7.3 upwards.

We're currently running 7.2.24 on our servers (hence the platform spec in `composer.json`), so only testing in 7.3 up would be risky.

0.3 is the latest version of WP_Mock which meets the PHP versions on our servers.

This commit also adds a call to WP_Mock's bootstrap() method, as per the documentation for v0.3.0